+++ examples/create_private_message.rb
@@ -5,13 +5,8 @@ client = DiscourseApi::Client.new("http://localhost:3000")
 client.api_key = "YOUR_API_KEY"
 client.api_username = "YOUR_USERNAME"
 
-# create user
-user = client.create_user(
-  name: "Bruce Wayne",
-  email: "bruce@wayne.com",
-  username: "batman",
-  password: "WhySoSerious"
+client.create_private_message(
+  title: "Confidential: Hello World!",
+  raw: "This is the raw markdown for my private message",
+  target_usernames: "user1,user2"
 )
-
-# activate user
-client.activate(user["user_id"])