+++ examples/sent_private_messages.rb
@@ -5,13 +5,4 @@ client = DiscourseApi::Client.new("http://localhost:3000")
 client.api_key = "YOUR_API_KEY"
 client.api_username = "YOUR_USERNAME"
 
-# create user
-user = client.create_user(
-  name: "Bruce Wayne",
-  email: "bruce@wayne.com",
-  username: "batman",
-  password: "WhySoSerious"
-)
-
-# activate user
-client.activate(user["user_id"])
+client.sent_private_messages('test_user')