+++ lib/discourse_api/api/posts.rb
@@ -4,6 +4,7 @@ module DiscourseApi
       def create_post(args)
         args = API.params(args)
                   .required(:topic_id, :raw)
+                  .optional(:created_at, :api_username)
         post("/posts", args)
       end
 
@@ -28,13 +29,16 @@ module DiscourseApi
         put("/posts/#{id}", post: {raw: raw})
       end
 
+      def delete_post(id)
+        delete("/posts/#{id}.json")
+      end
+
       def destroy_post_action(post_id, post_action_type_id)
         delete("/post_actions/#{post_id}.json", post_action_type_id: post_action_type_id)
       end
 
-      # This will need to be updated when Discourse 1.5 is released
       def post_action_users(post_id, post_action_type_id)
-        response = get("/post_actions/users.json", {id: post_id, post_action_type_id: post_action_type_id})
+        response = get("/post_action_users.json", {id: post_id, post_action_type_id: post_action_type_id})
         response[:body]
       end
     end