+++ lib/discourse_api/api/uploads.rb
@@ -1,0 +1,13 @@
+module DiscourseApi
+  module API
+    module Uploads
+      def upload_file(args)
+        args = API.params(args)
+                  .optional(:file, :url, :user_id)
+                  .default(type: 'composer', synchronous: true)
+                  .to_h
+        post('/uploads', args)
+      end
+    end
+  end
+end